go back to an address rtree layer
I guess we're back to an address rtree index, because overlapping rtree
layers just causes all sorts of problems. So instead of a rtree layer
per turf, have one big one but also tie each node [:WITHIN]-> its turf.

Oh and by the way we're also keeping the native point too because the
distance search there is way faster.

So, we have an address rtree layer, which requires the longitude/latitude
paramater on the address for the rtree index, which also creates a bbox
parameter on the address, we have a per-turf direct relationship for all
addresses, AND a native point() parameter for the distance search.

The same point information stored 4 different ways. What a waste of space!

Oh well, it's the trade-off for performance at scale.