Initial commit of Privacy Policy & Terms of Service
We commit these policies to source control to foster a culture of transparency. These policies are often quietly changed by organizations, which hide the details of when something changed, and why. We believe those details are important and stand by these terms as they are relevant today, and although this specific copy of these policies may be less relevant in the future, we intend to be open and honest about our history and how we are evolving as an organization.